Releases/3.0 - merge from master + bugfixes

+++ site/modules/slb/static/js/controllers/elb-list.js
@@ -3,8 +3,9 @@
 (function (app) {
     app.controller(
         'slb.ListController',
-        ['$scope', 'requestContext', 'localization', 'slb.Service', '$location', 'notification',
-                function ($scope, requestContext, localization, service, $location, notification) {
+        ['$scope', 'requestContext', 'localization', 'slb.Service', '$location', 'util',
+                function ($scope, requestContext, localization, service, $location, util) {
+
                 $scope.listLoaded = false;
                 localization.bind('slb', $scope);
                 requestContext.setUpRenderContext('slb.list', $scope, {
@@ -12,12 +13,24 @@
                 });
 
                 $scope.disableLb = function () {
-                    $scope.listLoaded = false;
-                    service.deleteController().then(function () {
-                        $location.path('/slb');
-                    }, function (err) {
-                        $scope.listLoaded = true;
-                    });
+                    util.confirm(
+                        localization.translate(
+                            $scope,
+                            null,
+                            'Confirm: Uninstall Load Balancer'
+                        ),
+                        localization.translate(
+                            $scope,
+                            null,
+                            'All load balancers will be deleted. This cannot be undone.'
+                        ), function () {
+                            $scope.listLoaded = false;
+                            service.deleteController().then(function () {
+                                $location.path('/slb');
+                            }, function () {
+                                $scope.listLoaded = true;
+                            });
+                        });
                 };
 
                 $scope.servers = [];
@@ -27,7 +40,19 @@
                         $scope.servers = data;
                         $scope.listLoaded = true;
                     }, function (err) {
-                        notification.replace('slb', { type: 'error' }, err);
+                        util.error(
+                            localization.translate(
+                                null,
+                                null,
+                                'Error'
+                            ),
+                            localization.translate(
+                                null,
+                                'slb',
+                                err
+                            ),
+                            function () {}
+                        );
                     });
                 }, function () {
                     $location.path('/slb');
